Remove unnecessary suppression of CC-RX's warning W0511174
This warning is the following.
"W0511174: The "-branch" option is ignored when the "-output=prep" option is specified at the same time."
At least CC-RX V3.02, specifying suppression of this warning is no more necessary.
And add necessary setting file created by e2 studio for CC-RX.

No changes are caused in the MOT and SREC files which are built.