Refactor Renesas RX72N Envision Kit RTOS Demo based on RISC-V RV32 SiFive HiFive1 RevB IAR RTOS Demo
Additionally,
(1) The simple blinky demo's 'Blink' messages are sent to not only the UART but also the Debug Console.
(2) The full demo sends '.' pass messages and/or various fail status messages to the Debug Console.
(3) The following error messages can be displayed in the Debug Console if it is detected.
Assertion failed!
Insufficient heap memory!
Stack Overflow!
(*) Note that EWRX desn't have the Debug Console.

Changes are caused in the MOT and SREC files which are built for Renesas RX72N Envision Kit RTOS Demo.